OpenCL version
The Qt version now has the integrator implemented in OpenCL using
boost.compute with the resulting values being directly copied to the
vertex buffer of the OpenGL widget. Performance is significantly
improved to the point where a 1,000,000 element simulation with 3 stages
in x and y can run in real time.